pyproject.toml
[project]
name = "fhamb-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aerich>=0.9.1",
    "aioboto3>=15.1.0",
    "aiosmtplib>=4.0.2",
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.117.1",
    "fitz>=0.0.1.dev2",
    "httpx>=0.28.1",
    "orjson>=3.11.3",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "python-slugify>=8.0.4",
    "redis>=6.4.0",
    "tortoise-orm[asyncpg]>=0.25.1",
    "uvicorn[standard]>=0.37.0",
]

[project.scripts]
dev = "src.scripts.server:run_server"
makemigrations = "src.scripts.aerich:makemigrations"
migrate = "src.scripts.aerich:migrate"
seed = "src.scripts.seed:run"

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src",

]

[tool.aerich]
tortoise_orm = "src.core.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
